requirements.txt
Flask==3.0.0
psycopg2==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.0
Flask-Login==0.6.3
requests==2.31.0
pytest==7.4.3
flake8==6.1.0
azure-cosmos==4.5.1